+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -11,6 +11,11 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Fields `meta: ListValueJSONObject!` and `metaKeys: [String!]!` for types `Comment/CustomPost/TaxonomyTerm/User` (#3060)
 - Type `ListValueJSONObject` (#3060)
 
+### Improvements
+
+- Made meta field `metaValue` handle any scalar type (previously only `String`) (#3061)
+- Made meta field `metaValues` handle any scalar type (previously only built-in ones), such as `JSONObject` (#3061)
+
 ## 11.2.0 - 15/03/2025
 
 ### Added